+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The status page now names the mechanism the server is waiting through - [#49](https://github.com/hivesolutions/viriatum/issues/49)
 * Test runs now report which tests left allocations behind them - [#49](https://github.com/hivesolutions/viriatum/issues/49)
 * Serving of a directory or of a python application straight from the command line - [#69](https://github.com/hivesolutions/viriatum/issues/69)
+* Naming of the configuration file to read, and a way of asking for none to be read at all - [#69](https://github.com/hivesolutions/viriatum/issues/69)
 
 ### Changed